Refactor MergeRequestManager, create base class in utils package
