+++ reconcile/aws_version_sync/integration.py
@@ -338,7 +338,7 @@ class AVSIntegration(QontractReconcileIntegration[AVSIntegrationParams]):
         external_resources_app_interface: AppInterfaceExternalResources,
     ) -> None:
         # initialize the merge request manager
-        merge_request_manager.fetch_avs_managed_open_merge_requests()
+        merge_request_manager._fetch_avs_managed_open_merge_requests()
         # housekeeping: close old/bad MRs
         merge_request_manager.housekeeping()
         diff = diff_iterables(
