+++ reconcile/aws_version_sync/merge_request_manager/merge_request_manager.py
@@ -1,9 +1,6 @@
 import logging
-import re
-from dataclasses import dataclass
 
 from gitlab.exceptions import GitlabGetError
-from gitlab.v4.objects import ProjectMergeRequest
 
 from reconcile.aws_version_sync.merge_request_manager.merge_request import (
     AVS_LABEL,
@@ -12,16 +9,13 @@ from reconcile.aws_version_sync.merge_request_manager.merge_request import (
     Renderer,
 )
 from reconcile.utils.gitlab_api import GitLabApi
-from reconcile.utils.merge_request_manager.parser import ParserError, ParserVersionError
+from reconcile.utils.merge_request_manager.merge_request_manager import (
+    MergeRequestManagerBase,
+    OpenMergeRequest,
+)
 from reconcile.utils.mr.base import MergeRequestBase
 from reconcile.utils.mr.labels import AUTO_MERGE
 from reconcile.utils.vcs import VCS
-
-
-@dataclass
-class OpenMergeRequest:
-    raw: ProjectMergeRequest
-    avs_info: AVSInfo
 
 
 class AVSMR(MergeRequestBase):
@@ -54,7 +48,7 @@ class AVSMR(MergeRequestBase):
         )
 
 
-class MergeRequestManager:
+class MergeRequestManager(MergeRequestManagerBase):
     """
     Manager for AVS merge requests. This class
     is responsible for housekeeping (closing old/bad MRs) and
@@ -68,92 +62,11 @@ class MergeRequestManager:
     def __init__(
         self, vcs: VCS, renderer: Renderer, parser: Parser, auto_merge_enabled: bool
     ):
-        self._vcs = vcs
+        super().__init__(vcs, parser, AVS_LABEL)
         self._open_mrs: list[OpenMergeRequest] = []
         self._open_mrs_with_problems: list[OpenMergeRequest] = []
-        self._open_raw_mrs: list[ProjectMergeRequest] = []
         self._renderer = renderer
-        self._parser = parser
         self._auto_merge_enabled = auto_merge_enabled
-
-    def _apply_regex(self, pattern: re.Pattern, promotion_data: str) -> str:
-        matches = pattern.search(promotion_data)
-        if not matches:
-            return ""
-        groups = matches.groups()
-        if len(groups) != 1:
-            return ""
-        return groups[0]
-
-    def fetch_avs_managed_open_merge_requests(self) -> None:
-        all_open_mrs = self._vcs.get_open_app_interface_merge_requests()
-        self._open_raw_mrs = [mr for mr in all_open_mrs if AVS_LABEL in mr.labels]
-
-    def housekeeping(self) -> None:
-        """
-        Close bad MRs:
-        - bad description format
-        - old AVS version
-        - merge conflict
-
-        --> if we bump the AVS version, we automatically close
-        old open MRs and replace them with new ones.
-        """
-        for mr in self._open_raw_mrs:
-            attrs = mr.attributes
-            desc = attrs.get("description")
-            has_conflicts = attrs.get("has_conflicts", False)
-            if has_conflicts:
-                logging.info(
-                    "Merge-conflict detected. Closing %s",
-                    mr.attributes.get("web_url", "NO_WEBURL"),
-                )
-                self._vcs.close_app_interface_mr(
-                    mr, "Closing this MR because of a merge-conflict."
-                )
-                continue
-            try:
-                avs_info = self._parser.parse(description=desc)
-            except ParserVersionError:
-                logging.info(
-                    "Old MR version detected! Closing %s",
-                    mr.attributes.get("web_url", "NO_WEBURL"),
-                )
-                self._vcs.close_app_interface_mr(
-                    mr, "Closing this MR because it has an outdated AVS version"
-                )
-                continue
-            except ParserError:
-                logging.info(
-                    "Bad MR description format. Closing %s",
-                    mr.attributes.get("web_url", "NO_WEBURL"),
-                )
-                self._vcs.close_app_interface_mr(
-                    mr, "Closing this MR because of bad description format."
-                )
-                continue
-
-            self._open_mrs.append(OpenMergeRequest(raw=mr, avs_info=avs_info))
-
-    def _merge_request_already_exists(
-        self,
-        provider: str,
-        account_id: str,
-        resource_provider: str,
-        resource_identifier: str,
-        resource_engine: str,
-    ) -> OpenMergeRequest | None:
-        for mr in self._open_mrs:
-            if (
-                mr.avs_info.provider == provider
-                and mr.avs_info.account_id == account_id
-                and mr.avs_info.resource_provider == resource_provider
-                and mr.avs_info.resource_identifier == resource_identifier
-                and mr.avs_info.resource_engine == resource_engine
-            ):
-                return mr
-
-        return None
 
     def create_avs_merge_request(
         self,
@@ -167,14 +80,15 @@ class MergeRequestManager:
         resource_engine_version: str,
     ) -> None:
         """Open new MR (if not already present) for an external resource and close any outdated before."""
-        if mr := self._merge_request_already_exists(
-            provider=provider,
-            account_id=provisioner_uid,
-            resource_provider=resource_provider,
-            resource_identifier=resource_identifier,
-            resource_engine=resource_engine,
-        ):
-            if mr.avs_info.resource_engine_version == resource_engine_version:
+        if mr := self._merge_request_already_exists({
+            "provider": provider,
+            "account_id": provisioner_uid,
+            "resource_provider": resource_provider,
+            "resource_identifier": resource_identifier,
+            "resource_engine": resource_engine,
+        }):
+            assert isinstance(mr.mr_info, AVSInfo)
+            if mr.mr_info.resource_engine_version == resource_engine_version:
                 # an MR for this external resource already exists
                 return None
             logging.info(
